requirements.txt
transformers
gradio
pandas
PyPDF2
ipykernel
spacy
torch
pinecone
python-dotenv
json5